refactor: put Search and PokemonList into components
