+++ src/reducers/index.ts
@@ -1,0 +1,10 @@
+// src/reducers/index.ts
+import { combineReducers } from 'redux';
+import pokemonReducer from './pokemonSlice';
+
+const rootReducer = combineReducers({
+  pokemon: pokemonReducer,
+  // Добавьте другие редьюсеры здесь
+});
+
+export default rootReducer;
