+++ src/store.js
@@ -1,4 +1,4 @@
-// store.js
+// src/store.ts
 import { configureStore } from '@reduxjs/toolkit';
 import rootReducer from './reducers';
 
@@ -7,3 +7,5 @@ const store = configureStore({
 });
 
 export default store;
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
